# openapi.yaml
openapi: 3.0.3
info:
  title: LeanStock Backend API
  description: |
    Production-grade inventory management system for coffee equipment retailers.
    Handles multi-tenant stock tracking, dead-stock decay pricing, demand forecasting,
    and full audit logging across warehouses.
  version: 1.0.0
  contact:
    name: LeanStock
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
          example: "eyJpZCI6ImFiYzEyMyJ9"
        hasMore:
          type: boolean
          example: true
        limit:
          type: integer
          example: 20
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request body
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Role:
      type: string
      enum: [ADMIN, MANAGER, STOREKEEPER]
    Category:
      type: string
      enum: [EQUIPMENT, COFFEE, ACCESSORIES]
    ProductStatus:
      type: string
      enum: [ACTIVE, DISPLAY_SAMPLE, DISCONTINUED, CRITICAL_DEADSTOCK]
    MovementType:
      type: string
      enum: [INCOME, EXPENSE, ADJUSTMENT, WRITEOFF]
...
    Tokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT, expires in 15 minutes
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: JWT, expires in 7 days, rotated on each use
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b1c2d3e4-f5a6-7890-abcd-ef1234567890"
        sku:
          type: string
          example: "JURA-E8-BLK"
        name:
          type: string
          example: "Jura E8 Black"
        brand:
          type: string
          example: "Jura"
        category:
          $ref: '#/components/schemas/Category'
        status:
          $ref: '#/components/schemas/ProductStatus'
        basePrice:
          type: integer
          description: Original price in KZT (tenge)
          example: 285000
        costPrice:
          type: integer
          description: Procurement cost in KZT
          example: 190000
        currentPrice:
          type: integer
          description: Effective selling price after any decay discount
          example: 256500
        discountPercent:
          type: integer
          description: Applied decay discount (0-50%)
          example: 10
        daysInInventory:
          type: integer
          description: Days since last stock replenishment
          example: 45
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Stock:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quantity:
          type: integer
          example: 7
        reorderThreshold:
          type: integer
          example: 5
        productId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
    MovementLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/MovementType'
        quantityDelta:
          type: integer
          description: Positive for INCOME, negative for EXPENSE/WRITEOFF
          example: -3
        quantityAfter:
          type: integer
          example: 7
        note:
          type: string
          nullable: true
          example: "Sold to customer #88"
        createdAt:
          type: string
          format: date-time
        stock:
          type: object
          properties:
            product:
              $ref: '#/components/schemas/Product'
            location:
              $ref: '#/components/schemas/Location'
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
            role:
              $ref: '#/components/schemas/Role'
    Location:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "d4e5f6a7-b8c9-0123-def0-1234567890ab"
        name:
          type: string
          example: "Almaty Warehouse"
        address:
          type: string
          example: "Abay Ave 150, Almaty"
        tenantId:
          type: string
          format: uuid
...
paths:
...
  /auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify email address via token link
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Token from the verification email
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email verified successfully. You can now log in."
        '400':
          description: Token missing, invalid, or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INVALID_TOKEN
                message: "Invalid or expired verification token"
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Issue new access + refresh token pair (rotates refresh token)
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tokens'
        '401':
          description: Invalid, expired, or revoked refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid:
                  value: { code: INVALID_REFRESH_TOKEN, message: "Invalid or expired refresh token" }
                revoked:
                  value: { code: TOKEN_REVOKED, message: "Refresh token has been revoked" }
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "manager@coffeeshop.kz"
      responses:
        '200':
          description: Always succeeds to prevent email enumeration
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If this email exists, a reset link has been sent."
        '429':
          description: Rate limit exceeded
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                  format: uuid
                newPassword:
                  type: string
                  minLength: 8
                  example: "NewSecurePass1"
      responses:
        '200':
          description: Password changed; all sessions revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successfully. You can now log in."
        '400':
          description: Token invalid or expired
...
  /products:
    get:
      tags: [Products]
      summary: List products with filters and cursor pagination
      operationId: listProducts
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/Category'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: brand
          in: query
          schema:
            type: string
            example: "Jura"
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, -createdAt, currentPrice, -currentPrice]
            default: -createdAt
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Not authenticated

    post:
      tags: [Products]
      summary: Create product (MANAGER+)
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sku, name, brand, category, basePrice, costPrice]
              properties:
                sku:
                  type: string
                  pattern: '^[A-Z0-9-]+$'
                  minLength: 4
                  maxLength: 20
                  example: "JURA-E8-BLK"
                name:
                  type: string
                  example: "Jura E8 Black"
                brand:
                  type: string
                  example: "Jura"
                category:
                  $ref: '#/components/schemas/Category'
                basePrice:
                  type: integer
                  minimum: 0
                  example: 285000
                costPrice:
                  type: integer
                  minimum: 0
                  example: 190000
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          description: Requires MANAGER or ADMIN role
        '409':
          description: SKU already exists
  /products/{id}:
    get:
      tags: [Products]
      summary: Get single product
      operationId: getProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found

    patch:
      tags: [Products]
      summary: Update product fields (MANAGER+)
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                brand:
                  type: string
                basePrice:
                  type: integer
                  minimum: 0
                status:
                  type: string
                  enum: [ACTIVE, DISPLAY_SAMPLE, DISCONTINUED]
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    delete:
      tags: [Products]
      summary: Archive product — sets status to DISCONTINUED (MANAGER+)
      operationId: archiveProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product archived
  /inventory/income:
    post:
      tags: [Inventory]
      summary: Receive goods into a location (all roles)
      operationId: recordIncome
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, locationId, quantity]
              properties:
                productId:
                  type: string
                  format: uuid
                locationId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
                  example: 10
                note:
                  type: string
                  example: "Delivery from supplier"
      responses:
        '201':
          description: Stock updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        '404':
          description: Product or location not found
  /inventory/expense:
    post:
      tags: [Inventory]
      summary: Record goods leaving a location (sale, damage, write-off)
      operationId: recordExpense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, locationId, quantity, reason]
              properties:
                productId:
                  type: string
                  format: uuid
                locationId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
                  example: 3
                reason:
                  type: string
                  enum: [SALE, DAMAGE, RETURN, WRITEOFF]
                  example: SALE
                note:
                  type: string
                  example: "Customer sale #1042"
      responses:
        '201':
          description: Stock updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        '404':
          description: No stock record at this location
        '409':
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: INSUFFICIENT_STOCK
                message: "Insufficient stock: requested 5, available 3"
                details:
                  requested: 5
                  available: 3
  /inventory/transfer:
    post:
      tags: [Inventory]
      summary: Transfer stock between locations atomically (MANAGER+)
      operationId: transferStock
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, fromLocationId, toLocationId, quantity]
              properties:
                productId:
                  type: string
                  format: uuid
                fromLocationId:
                  type: string
                  format: uuid
                toLocationId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
                  example: 4
                note:
                  type: string
      responses:
        '200':
          description: Transfer complete — both location records updated atomically
          content:
            application/json:
              schema:
                type: object
                properties:
                  from:
                    $ref: '#/components/schemas/Stock'
                  to:
                    $ref: '#/components/schemas/Stock'
        '400':
          description: Source and destination are the same location
        '404':
          description: No stock at source location
        '409':
          description: Insufficient stock at source
  /inventory/history:
    get:
      tags: [Inventory]
      summary: Movement history with cursor pagination
      operationId: getMovementHistory
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
        - name: productId
          in: query
          schema:
            type: string
            format: uuid
        - name: locationId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/MovementType'
      responses:
        '200':
          description: Paginated movement log
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovementLog'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...